Refactored CertID validation to separate package/class
